+++ src/admin-scripts/importEventsDefinition.ts
@@ -2,7 +2,7 @@ import log4js from 'log4js'
 import envCheck from '../envCheck'
 import container from '../di/inversify.config'
 import { types } from '../di/types'
-import { UpdateSchoolCalenderUseCase } from '../usecase/UpdateSchoolCalenderUseCase'
+import { UpdateSchoolCalenderUseCase } from '../usecase/updateSchoolCalenderUseCase'
 import { promises as fs } from 'fs'
 import { connect } from '../infrastructure/database'
