+++ src/di/inversify.config.ts
@@ -4,11 +4,11 @@ import { types } from './types'
 import { PLectureRepository } from '../infrastructure/database/pLectureRepository'
 import { RemoteLectureRepository } from '../interface/repository/remoteLectureRepository'
 import { KdbRemoteLectureRepository } from '../infrastructure/database/kdbRemoteLectureRepository'
-import { UpdateLectureDatabaseUseCase } from '../usecase/UpdateLectureDatabaseUseCase'
-import { UpdateLectureDatabaseInteractor } from '../interactor/UpdateLectureDatabaseInteractor'
+import { UpdateLectureDatabaseUseCase } from '../usecase/updateLectureDatabaseUseCase'
+import { UpdateLectureDatabaseInteractor } from '../interactor/updateLectureDatabaseInteractor'
 import { LectureController } from '../interface/controller/lectureController'
-import { FindLectureUseCase } from '../usecase/FindLectureUseCase'
-import { FindLectureInteractor } from '../interactor/FindLectureInteractor'
+import { FindLectureUseCase } from '../usecase/findLectureUseCase'
+import { FindLectureInteractor } from '../interactor/findLectureInteractor'
 import { UserRepository } from '../interface/repository/userRepository'
 import { PUserRepository } from '../infrastructure/database/pUserRepository'
 import { CreateUserUseCase } from '../usecase/createUserUseCase'
@@ -34,11 +34,11 @@ import { UserLectureController } from '../interface/controller/userLectureContro
 import { FindUserLectureUseCase } from '../usecase/findUserLectureUseCase'
 import { FindUserLectureInteractor } from '../interactor/findUserLectureInteractor'
 import { UpdateUserLectureUseCase } from '../usecase/updateUserLectureUseCase'
-import { UpdateUserLectureInteractor } from '../interactor/UpdateUserLectureInteractor'
+import { UpdateUserLectureInteractor } from '../interactor/updateUserLectureInteractor'
 import { SchoolCalenderRepository } from '../interface/repository/schoolCalenderRepository'
 import { PSchoolCalenderRepository } from '../infrastructure/database/pSchoolCalenderRepository'
-import { UpdateSchoolCalenderUseCase } from '../usecase/UpdateSchoolCalenderUseCase'
-import { UpdateSchoolCalenderInteractor } from '../interactor/UpdateSchoolCalenderInteractor'
+import { UpdateSchoolCalenderUseCase } from '../usecase/updateSchoolCalenderUseCase'
+import { UpdateSchoolCalenderInteractor } from '../interactor/updateSchoolCalenderInteractor'
 import { GetSchoolCalenderUseCase } from '../usecase/getSchoolCalenderUseCase'
 import { GetSchoolCalenderInteractor } from '../interactor/getSchoolCalenderInteractor'
 import { SchoolCalenderController } from '../interface/controller/schoolCalenderController'
