+++ src/usecase/findLectureUseCase.ts
