+++ src/usecase/updateLectureDatabaseUseCase.ts
